Handle arbitrary custom events
By letting the `process-event` function dynamically create the keyword
where the event handlers for this type of event are stored, we can
allow a game to define their own custom events, and dispatch them with
`enqueue-event!`.

Could be used for complex event handling, or even multiplayer
stuff (with core.async?).