Preload assets
We don't want to load images and audio files (especially large music
files) at runtime, so we're a little preloader with a progress bar
during the game init.

Assets are defined in the game config, and can be referenced by their
key.

To make it easy to refer to loaded assets we've created an atom (map)
to hold image textures mirroring how we store audio buffers.

As part of this (to draw the progress bar) we've added a `clunk.shape`
namespace for drawing simple shapes.