Support concave polygons through triangulation.
We've added a new `clunk.shape/fill-concave-poly!` function which
takes a poly and splits it into triangles, then draws the
triangles. Decide to have it separate to the original function as it's
more expensive.

Alternatively we could just use triangulate when creating sprites and
then draw the triangles manually? Use choice.